+++ matrixchain.cpp
@@ -15,8 +15,7 @@ int main(int argc, char** argv) {
         std::cin >> rows;
         mc.append(Matrix<int>(cols, rows));
     }
-
-    std::cout << mc.naive_multiply_cost() / static_cast<double>(mc.matrix_chain_order());
+    std::cout << mc.naive_multiply_cost() / static_cast<long double>(mc.matrix_chain_order());
 
     return 0;
 }
